design_assistant/requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.3
streamlit==1.41.1